requirements.txt
flask
python-dotenv
requests
openai
python-dotenv